requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.5
fastapi==0.115.0
idna==3.11
pydantic==2.9.2
python-dotenv==1.2.2
requests==2.32.5
urllib3==2.6.3
uvicorn==0.30.6
